Independent SCP-1162 entity
This commit adds the new SCP-1162 object, and removes SCP-1162 hole from Room_1162.
The functions and such are moved entirely to anomaly folder.
