+++ addons/anomaly/CfgFunctions.hpp
@@ -1,0 +1,7 @@
+class CfgFunctions {
+	class SCP {
+		class Functions {
+			class 1162_itemExchange { file = QPATHTOF(functions\fnc_1162_itemExchange.sqf); };
+		};
+	};
+};
