+++ addons/anomaly/CfgPatches.hpp
@@ -13,7 +13,8 @@ class CfgPatches
             "SCP_CB_Anomaly_SCP079_X",
             "SCP_CB_Anomaly_SCP106",
             "SCP_CB_Anomaly_SCP173_Static",
-            "SCP_CB_Anomaly_SCP294"
+            "SCP_CB_Anomaly_SCP294",
+            "SCP_CB_Anomaly_SCP1162"
         };
         weapons[] =
         {
