+++ addons/anomaly/data/model.cfg
@@ -192,4 +192,9 @@ class CfgModels
 		sectionsInherit = "";
 		sections[] = {"camo"};
 	};
+
+	class SCP1162 {
+		sectionsInherit = "";
+		sections[] = {"camo"};
+	};
 };
