+++ addons/anomaly/stringtable.xml
@@ -53,5 +53,13 @@
             <English>SCP-294 "The Coffee Machine"</English>
             <Japanese>構造物（SCP）</Japanese>
         </Key>
+        <Key ID="STR_RegalisSCP_CB_Anomaly_SCP1162">
+            <English>SCP-1162 "A Hole in the Wall"</English>
+            <Japanese>構造物（SCP）</Japanese>
+        </Key>
+        <Key ID="STR_RegalisSCP_CB_Anomaly_SCP1162_ReachInto">
+            <English>Reach into the hole</English>
+            <Japanese>穴の中に手を伸ばす</Japanese>
+        </Key>
     </Package>
 </Project>
