--- addons/room_lcz/CfgFunctions.hpp
@@ -1,7 +1,0 @@
-class CfgFunctions {
-	class SCP {
-		class Functions {
-			class 1162_itemExchange { file = QPATHTOF(functions\fnc_1162_itemExchange.sqf); };
-		};
-	};
-};
