+++ addons/room_lcz/CfgVehicles.hpp
@@ -721,23 +721,6 @@ class CfgVehicles
         displayName = CSTRING(Room_1162);
         model = QPATHTOF(data\Room_1162\Room_1162.p3d);
         icon = "iconObject_1x1";
-		class UserActions
-		{
-			class SCP1162
-			{
-				displayNameDefault = "<img image='z\RegalisSCP_CB\addons\textures\handsymbol.paa' size='2.5' />";
-				displayName = CSTRING(Room_1162_ReachInto);
-                priority = 6;
-				position = "scp_1162";
-				radius = 2;
-				aiMaxRange = 2;
-				onlyForPlayer = 0;
-				condition = "alive this";
-				//statement = (this execVM ("z\regalisscp_cb\addons\room_lcz\data\Room_1162\1162_script.sqf"));
-                statement = "[this] call SCP_fnc_1162_itemExchange;";
-                
-			};
-		};
 	};
 //******************//
 // Hallways & Rooms //
