--- addons/room_lcz/functions/script_component.hpp
@@ -1,1 +1,0 @@
-#include "\z\RegalisSCP_CB\addons\room_lcz\script_component.hpp"
